+++ integration/tarantoolee_gost_pure_test.go
@@ -125,7 +125,7 @@ func TestTarantoolEE_Ping_GOST_Pure(t *testing.T) {
 			require.NoErrorf(t, waitForTCP(addr, 30*time.Second),
 				"tarantool-ee did not open %s", addr)
 
-			dialer := tlsdialer.OpenSSLDialer{
+			dialer := tlsdialer.TLSDialer{
 				Address:    fmt.Sprintf("localhost:%d", port),
 				User:       "test",
 				Password:   "test",