+++ integration/tarantoolee_gost_rsa_clientauth_test.go
@@ -100,7 +100,7 @@ func TestTarantoolEE_Ping_GOSTServer_RSAClientAuth(t *testing.T) {
 	require.NoErrorf(t, waitForTCP(addr, 30*time.Second),
 		"tarantool-ee did not open %s", addr)
 
-	dialer := tlsdialer.OpenSSLDialer{
+	dialer := tlsdialer.TLSDialer{
 		Address:     fmt.Sprintf("localhost:%d", port),
 		User:        "test",
 		Password:    "test",